MAINT: move dict invert to functional and remove unused invert_unique
